Move DynamicQuantities extension to within package
