+++ src/SymbolicRegression.jl
@@ -141,6 +141,7 @@ include("RegularizedEvolution.jl")
 include("SingleIteration.jl")
 include("ProgressBars.jl")
 include("Migration.jl")
+include("DimensionalAnalysis.jl")
 include("SearchUtils.jl")
 
 import .CoreModule:
@@ -1015,7 +1016,6 @@ end
 #! format: off
 if !isdefined(Base, :get_extension)
     @init @require SymbolicUtils = "d1185830-fcd6-423d-90d6-eec64667417b" include("../ext/SymbolicRegressionSymbolicUtilsExt.jl")
-    @init @require DynamicQuantities = "06fc5a27-2a28-4c7c-a15d-362465fb6821" include("../ext/SymbolicRegressionDynamicQuantitiesExt.jl")
 end
 #! format: on
